+++ src/models/accinfo.ts
@@ -3,9 +3,14 @@ export interface IAccountInfo {
     accounts: IAccount[];
 }
 export interface IAccount {
+    alias: string;
     guid: string;
     password: string;
     serverPref: string;
+    charInfo: ICharacterInfo;
+}
+
+export interface ICharacterInfo {
     charId: number;
     nextCharId: number;
     maxNumChars: number;